+++ schnose_discord/src/formatting.rs
@@ -4,13 +4,13 @@
 /// let time: f32 = 1263.7832;
 /// assert_eq!("00:21:03.78", &format_time(time));
 /// ```
-pub(crate) fn format_time(secs_float: f32) -> String {
+pub(crate) fn format_time(secs_float: f64) -> String {
 	let seconds = secs_float as u32;
 	let hours = ((seconds / 3600) % 24) as u8;
 	let seconds = seconds % 3600;
 	let minutes = (seconds / 60) as u8;
 	let seconds = seconds % 60;
-	let millis = ((secs_float - (secs_float as u32) as f32) * 1000.0) as u16;
+	let millis = ((secs_float - (secs_float as u32) as f64) * 1000.0) as u16;
 
 	let mut s = format!("{:02}:{:02}.{:03}", minutes, seconds, millis);
